Put Version class into it's own file.
