+++ lib/puppet/provider/php_version/php_source.rb
@@ -1,19 +1,5 @@
 require 'puppet/util/execution'
-
-class Version < Array
-  def initialize s
-    super(s.split('.').map { |e| e.to_i })
-  end
-  def < x
-    (self <=> x) < 0
-  end
-  def > x
-    (self <=> x) > 0
-  end
-  def == x
-    (self <=> x) == 0
-  end
-end
+require 'lib/util/version'
 
 Puppet::Type.type(:php_version).provide(:php_source) do
   include Puppet::Util::Execution
