+++ lib/util/version.rb
@@ -1,0 +1,14 @@
+class Version < Array
+  def initialize s
+    super(s.split('.').map { |e| e.to_i })
+  end
+  def < x
+    (self <=> x) < 0
+  end
+  def > x
+    (self <=> x) > 0
+  end
+  def == x
+    (self <=> x) == 0
+  end
+end
